pyproject.toml
[project]
name = "airflow-jaffle-shop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-storage-blob>=12.25.1",
    "duckdb>=1.2.2",
    "pandas>=2.2.3",
]

[dependency-groups]
dev = [
    "apache-airflow>=3.0.0",
]
